requirements.txt
fastapi
uvicorn
pydantic
aiohttp
slowapi
apitally[fastapi]
python-dotenv